refactor : 조 현황 화면 GridView 전환 및 파일 분리
- Wrap+LayoutBuilder로 손수 계산하던 카드 그리드 폭을
  GridView.builder + SliverGridDelegateWithMaxCrossAxisExtent로 교체
  (플랫폼이 이미 제공하는 max-extent 그리드 레이아웃, ponytail-audit 지적)
- group_list_screen.dart 하나에 몰려 있던 화면(GroupListScreen) +
  카드 위젯(_GroupStatusCard) + 조 등록 다이얼로그(_RegisterGroupDialog,
  QR 스캔/배지 목록/직접 입력 3-tab)를 dashboard 등 다른 feature와
  같은 컨벤션으로 위젯 파일 분리(_group_status_card.dart,
  _register_group_dialog.dart)
